+++ plugin/app/Services/Flows/SaveTesterCallback.php
@@ -112,6 +112,9 @@ class SaveTesterCallback
      */
     private function executeSave(TesterCallbackRequest $request, GitCallback $gitCallback, array $users, int $courseId = null): Submission
     {
+        global $CFG;
+        require_once ($CFG->dirroot . '/mod/charon/lib.php');
+
         $submission = $this->createNewSubmission($request, $gitCallback, $users[0]->id, $courseId);
 
         if ($request['files']) {
@@ -129,6 +132,10 @@ class SaveTesterCallback
         $this->charonGradingService->calculateCalculatedResultsForNewSubmission($submission);
 
         $this->updateGrades($submission, $users);
+
+        foreach ($users as $student) {
+            update_charon_completion_state($submission, $student->id);
+        }
 
         return $submission;
     }