+++ plugin/app/Services/ReviewCommentService.php
@@ -2,6 +2,7 @@
 
 namespace TTU\Charon\Services;
 
+use TTU\Charon\Exceptions\ReviewCommentException;
 use TTU\Charon\Repositories\ReviewCommentRepository;
 use Zeizig\Moodle\Globals\User;
 
@@ -30,40 +31,40 @@ class ReviewCommentService
      *
      * @param $submissionFileId
      * @param $reviewComment
-     * @return string[]
+     * @param $notify
+     * @throws ReviewCommentException
      */
-    public function add($submissionFileId, $reviewComment): array
+    public function add($submissionFileId, $reviewComment, $notify): void
     {
+        if (strlen($reviewComment) > 10000) {
+            throw new ReviewCommentException("review_comment_over_limit");
+        }
         $userId = app(User::class)->currentUserId();
-        $this->reviewCommentRepository->add($userId, $submissionFileId, $reviewComment);
-        return [
-            'status'  => 'OK'
-        ];
+        $this->reviewCommentRepository->add($userId, $submissionFileId, $reviewComment, $notify);
     }
 
     /**
      * Delete review comment.
      *
      * @param $reviewCommentId
-     * @return array
+     * @throws ReviewCommentException
      */
-    public function delete($reviewCommentId): array
+    public function delete($reviewCommentId): void
     {
         $comment = $this->reviewCommentRepository->get($reviewCommentId);
-        if ($comment) {
-            $result = $this->reviewCommentRepository->delete($reviewCommentId);
-            if ($result) {
-                return [
-                    'status' => 'OK'
-                ];
-            } else {
-                return [
-                    'status'=>'Failed',
-                ];
-            }
+        if (!$comment) {
+            throw new ReviewCommentException("delete_review_comment_not_found");
         }
-        return [
-            'status'=>'Failed',
-        ];
+        $this->reviewCommentRepository->delete($reviewCommentId);
+    }
+
+    /**
+     * Remove notification setting from review comments got by given identifiers.
+     *
+     * @param $reviewCommentIds
+     */
+    public function clearNotifications($reviewCommentIds): void
+    {
+        $this->reviewCommentRepository->clearNotification($reviewCommentIds);
     }
 }