+++ plugin/resources/assets/js/api/ReviewComment.js
@@ -1,23 +1,42 @@
 class ReviewComment {
 
-    static add(reviewComment, submissionFileId, charonId, then) {
+    static add(reviewComment, submissionFileId, charonId, notify, then) {
         axios.post('/mod/charon/api/charons/' + charonId + '/reviewComments/add', {
             submission_file_id: submissionFileId,
-            review_comment: reviewComment
+            review_comment: reviewComment,
+            notify: notify,
         }).then(response => {
             then(response.data)
         }).catch(error => {
-            VueEvent.$emit('show-notification', 'Error adding review comment.\n' + error, 'danger')
+            this.throwError(error, 'Error adding review comment.\n');
         })
+
     }
 
     static delete(reviewCommentId, charonId, then) {
-        axios.delete('/mod/charon/api/charons/' + charonId + '/reviewComments/' + reviewCommentId + '/delete')
-            .then(response => {
-                then(response.data)
-            }).catch(error => {
-            VueEvent.$emit('show-notification', 'Error deleting review comment.\n' + error, 'danger')
+        axios.delete('/mod/charon/api/charons/' + charonId + '/reviewComments/' + reviewCommentId + '/delete'
+        ).then(response => {
+            then(response.data)
+        }).catch(error => {
+            this.throwError(error, 'Error deleting review comment.\n');
         });
+    }
+
+    static clearNotifications(reviewCommentIds, charonId, studentId, then) {
+        axios.put('/mod/charon/api/charons/' + charonId + '/reviewComments/clear?user_id=' + studentId, {
+            reviewCommentIds: reviewCommentIds,
+        }).then(response => {
+            then(response.data)
+        }).catch(error => {
+            this.throwError(error, 'Error clearing review comments\' notifications.\n');
+        });
+    }
+
+    static throwError(error, errorText) {
+        VueEvent.$emit('show-notification',
+            error.response && error.response.data && error.response.data.title
+                ? error.response.data.title + ' ' + error.response.data.detail
+                : errorText + error, 'danger')
     }
 }