+++ plugin/resources/assets/js/components/partials/FilesComponent.vue
@@ -1,17 +1,5 @@
 <template>
     <v-card class="mx-auto" max-width="70vw" outlined raised>
-
-        <v-card>
-            <v-container v-if="activeFileId" class="gray-part">
-                <textarea rows="8" type="text" class="review-comment" v-model="newReviewComment" maxlength="10000"
-                          placeholder="Write a comment for the selected code (visible for the student)">
-                </textarea>
-                <v-btn class="review-comment-button ma-2" tile outlined color="primary"
-                       :disabled="!newReviewComment" @click="addReviewComment">
-                    Add comment
-                </v-btn>
-            </v-container>
-        </v-card>
         <div
                 class="columns is-gapless code-container"
                 :class="{ 'is-round': isRound }"
@@ -40,6 +28,18 @@
                 <pre class="code" v-highlightjs="activeFile.contents"><code :class="testerType"></code></pre>
             </div>
         </div>
+
+        <v-container v-if="activeFileId" class="gray-part">
+                <textarea rows="8" type="text" class="review-comment" v-model="newReviewComment" maxlength="10000"
+                          placeholder="Write a comment for the selected code (visible for the student)">
+                </textarea>
+            <v-btn class="review-comment-button ma-2" tile outlined color="primary"
+                   :disabled="!newReviewComment" @click="addReviewComment">
+                Add comment
+            </v-btn>
+            <input type="checkbox" class="form-control" v-model="notify">
+            Notify the student about the comment
+        </v-container>
     </v-card>
 
 </template>
@@ -47,7 +47,7 @@
 <script>
 
     import FileTree from './FileTree'
-    import ReviewComment from "../../api/ReviewComment";
+    import {ReviewComment} from "../../api";
     import {mapState} from "vuex";
 
     export default {
@@ -69,6 +69,7 @@
                 activeFileId: null,
                 formattedFiles: [],
                 newReviewComment: '',
+                notify: true
             }
         },
 
@@ -210,7 +211,7 @@
                     return
                 }
 
-                ReviewComment.add(this.newReviewComment.trim(), this.activeFileId, this.charon.id, () => {
+                ReviewComment.add(this.newReviewComment.trim(), this.activeFileId, this.charon.id, this.notify, () => {
                     this.newReviewComment = ''
                     VueEvent.$emit('show-notification', 'Review comment added!')
                     this.$root.$emit('refresh_submission_files')
@@ -313,6 +314,10 @@
 
     .gray-part {
         background-color: darken(#fafafa, 5%);
+    }
+
+    .is-gapless.code-container {
+        margin-bottom: 0;
     }
 
 </style>